hexToRGBA - rename hexToRGB to getRGBvalues in hextorgb package - update hexToRGB to accept one colour value arg - create new hexToRGBA package - update tests for both hexToRGB and hexToRGBA
