+++ packages/hextorgb/README.md
@@ -1,11 +1,3 @@
 # `@artifak/hextorgb`
 
-A utility function to convert hexadecimal colour values to RGB or RGBA string.
-
-## Usage
-
-```
-const hextorgb = require('@artifak/hextorgb');
-
-// TODO: DEMONSTRATE API
-```
+A utility function to convert hexadecimal colour values to RGB string.
