+++ packages/hextorgba/README.md
@@ -1,0 +1,3 @@
+# `@artifak/hextorgba`
+
+A utility function to convert hexadecimal colour values to RGBA string.
