+++ packages/hextorgba/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "@artifak/hextorgb",
+  "name": "@artifak/hextorgba",
   "version": "1.0.0",
-  "description": "Utility function to convert style hex colour codes to RGB string.",
+  "description": "Utility function to convert style hex colour codes to RGBA string.",
   "keywords": [
     "color",
-    "RGB",
+    "RGBA",
     "hexadecimal",
     "utilities"
   ],
   "author": "Julian Low",
   "license": "MIT",
   "sideEffects": false,
-  "main": "dist/hextorgb.cjs.js",
-  "module": "dist/hextorgb.esm.js",
+  "main": "dist/hextorgba.cjs.js",
+  "module": "dist/hextorgba.esm.js",
   "src": "src/index.ts",
   "types": "dist/src/index.d.ts",
   "directories": {
@@ -39,5 +39,8 @@
   "homepage": "https://github.com/heyjul3s/artifak#readme",
   "devDependencies": {
     "@artifak/bundler": "^1.1.3"
+  },
+  "dependencies": {
+    "@artifak/hextorgb": "^1.0.0"
   }
 }
