+++ packages/hextorgba/src/index.ts
@@ -1,0 +1,2 @@
+export { hexToRGBA } from './hextorgba';
+export { getRGBvalues } from '@artifak/hextorgb';
